fixed sync on deleted items

+++ src/app/pages/items/sync.service.ts
@@ -69,12 +69,19 @@ export class SyncService {
 
     for (const localItem of localItems) {
       if (localItem.deleted_at) {
+        let remoteDeleted = true;
         if (localItem.remote_id) {
           this.apiChanges = true;
-          await this.apiItemService.destroy(localItem.remote_id).toPromise();
+          try {
+            await this.apiItemService.destroy(localItem.remote_id).toPromise();
+          } catch (error) {
+            remoteDeleted = false;
+          }
         }
-        this.dbChanges = true;
-        await this.dbItemService.forceDelete(localItem.id);
+        if (remoteDeleted) {
+          this.dbChanges = true;
+          await this.dbItemService.forceDelete(localItem.id);
+        }
         continue;
       }
 
@@ -101,6 +108,10 @@ export class SyncService {
             .update(localItem.remote_id, localItem)
             .toPromise();
         }
+        if (!remoteItem) {
+          this.dbChanges = true;
+          await this.dbItemService.forceDelete(localItem.id);
+        }
         continue;
       }
     }
@@ -123,10 +134,13 @@ export class SyncService {
       localItem.remote_id = localItem.id;
       localItem.shopping_list_id = this.shoppingList.id;
       localItem.id = null;
+
       if (!found) {
         this.dbChanges = true;
         await this.dbItemService.insert(localItem);
-      } else if (remoteItem.updated_at > found.updated_at) {
+      }
+
+      if (found && remoteItem.updated_at > found.updated_at) {
         this.dbChanges = true;
         await this.dbItemService.update(found.id, localItem);
       }
